remove firebase from build
