+++ src/front/shared/containers/App/App.tsx
@@ -5,7 +5,11 @@ import PropTypes from "prop-types";
 import actions from "redux/actions";
 import { connect } from "redaction";
 import moment from "moment-with-locales-es6";
-import { constants, localStorage, firebase } from "helpers";
+import {
+  constants,
+  localStorage,
+  // firebase
+} from "helpers";
 import { isMobile } from "react-device-detect";
 
 import CSSModules from "react-css-modules";
@@ -309,7 +313,7 @@ class App extends React.Component<RouteComponentProps<any>, any> {
       }
     }
 
-    firebase.initialize();
+    // firebase.initialize();
 
     this.processUserBackup()
     this.processMetamask()
@@ -351,7 +355,7 @@ class App extends React.Component<RouteComponentProps<any>, any> {
   componentDidUpdate() {
     this.checkIfDashboardModalsAllowed()
     if (process.env.MAINNET) {
-      firebase.setUserLastOnline();
+      // firebase.setUserLastOnline();
     }
   }
 
@@ -416,7 +420,7 @@ class App extends React.Component<RouteComponentProps<any>, any> {
     const isWidgetBuild = config && config.isWidget;
 
     if (isWidgetBuild && localStorage.getItem(constants.localStorage.didWidgetsDataSend) !== "true") {
-      firebase.submitUserDataWidget("usersData");
+      // firebase.submitUserDataWidget("usersData");
       localStorage.setItem(constants.localStorage.didWidgetsDataSend, true);
     }
