+++ src/front/shared/pages/Wallet/Wallet.tsx
@@ -7,11 +7,16 @@ import cssModules from 'react-css-modules'
 import styles from './Wallet.scss'
 import { isMobile } from 'react-device-detect'
 import moment from 'moment'
-import firestore from 'helpers/firebase/firestore'
+// import firestore from 'helpers/firebase/firestore'
 
 import History from 'pages/History/History'
 
-import helpers, { firebase, links, constants, stats } from 'helpers'
+import helpers, {
+  // firebase,
+  links,
+  constants,
+  stats
+} from 'helpers'
 import { localisedUrl } from 'helpers/locale'
 import { getActivatedCurrencies } from 'helpers/user'
 import getTopLocation from 'helpers/getTopLocation'
@@ -437,7 +442,7 @@ export default class Wallet extends Component<any, any> {
       if (isOneHourAfter || isFirstCheck) {
         localStorage.setItem(constants.localStorage.lastCheckBalance, now)
         try {
-          const ipInfo = await firebase.getIPInfo()
+          const ipInfo = await stats.getIPInfo()
 
           const registrationData = {
             locale:
@@ -475,7 +480,7 @@ export default class Wallet extends Component<any, any> {
 
           await stats.updateUser(ethData.address, getTopLocation().host, registrationData)
 
-          firestore.updateUserData(balancesData)
+          // firestore.updateUserData(balancesData)
         } catch (error) {
           console.error(`Sync error in wallet: ${error}`)
         }
