+++ src/front/shared/redux/actions/user.ts
@@ -668,10 +668,10 @@ const getAuthData = (name) => {
   return user[`${name}Data`]
 }
 
-const addMessagingToken = (token) => {
-  console.log("Added firebase token to redux store: ", token)
-  reducers.user.addMessagingToken({ token })
-}
+// const addMessagingToken = (token) => {
+//   console.log("Added firebase token to redux store: ", token)
+//   reducers.user.addMessagingToken({ token })
+// }
 
 export default {
   sign,
@@ -692,5 +692,5 @@ export default {
   getWithdrawWallet,
   fetchMultisigStatus,
   pullActiveCurrency,
-  addMessagingToken,
+  // addMessagingToken,
 }
